Initial commit of a plug-in project
This Eclipse plug-in provides basic syntax highlighting for .jl editors,
i.e. for files containing Dolmen lexer descriptions. It is based on the
example editor plug-in from Eclipse, so there's is still a dummy icon
and some useless stuff around. But it basically works! In particular it
recognizes Java actions during partitioning so we can have correct Java
syntax highlihgting in semantic actions.