Added handlers to manage the Dolmen nature, and fill .jl builder command
At the moment, the lexer builder seems not to be called at all, even if
it is added to the builder list when the Dolmen nature is added. I must
investigate. Also, the Add/Remove nature commands should only be
available when applicable but the projectNature propertyTester does not
seem to work.