Ignore output folders when building resources
And also all files which are not in classpath entries.
This removes an issue where lexers/grammars were apparently built twice, whereas what happened was simply that the JDT copies them as is to the /bin/ folder, and they are thus passed in the ResourceDeltas as any other added/changed files.