Start using IDocumentSetupParticipant extensions to set up partitioners instead of specific DocumentProviders
Before, partitioners were connected to .jl and .jg documents in their respective DocumentProviders which extended FileDocumentProvider. Unfortunately, it offers limited support for undo/redo operations: in particular the dirty flag was not cleared when undoing all modifs in a buffer. Using TextFileDocumentProvider instead solved this issue, but was losing the ability to manually install the adequate partitioners.
We now use default TextFileDocumentProviders but use two IDocumentSetupParticipant extensions which are called when  setting up a Dolmen document, and use them to install the correct partitioners. Everything seems to work fine with this solution.