Adding an Outline view to .jl editors
To that end, JLEditor keeps an up-to-date lexer by parsing the document each time it is saved. The outline of the parsed lexer is displayed in the editor's Outline view, or the issue reports if the lexer could not be parsed.

Selection of nodes in the Outline view highlights/selects/reveals the corresponding piece of the editor when possible.