Adding an Outline to .jg editors
Like JLEditor, JGEditor now parses its document into a Grammar description each time it is saved. The grammar is shown in the Outline view, or the issue reports if the grammar could not be parsed.

Selection in the Outline highlights/selects/reveals the corresponding part in the editor if possible.