Improve partitioner for semantic actions in Dolmen documents
Defined an extension of RuleBasedPartitionScanner which prevents resuming scan from inside a partition (not suitable when parsing nested delimiters), and made JavaActionRule stop at closest most-shallow closing delimiter when a complete Java action is not found. This works very well in practice: partitions tend to stay 'correct' when typing Java code and opening/closing braces on the fly.