Add options partitioning and scanning to lexer editors
Option partinioning was easy in parser editors because grammar
descriptions do not use '[' and ']' other than in options.
In lexers, wa have to design our own rule-based scanner matching
[ <ident> = <some string> ].
We use it in both lexer and parser editors for the sake of uniformity
although it might be slightly less efficient in parsers than the former
rule. We also share the styling scanner for option partitions.

NB: for now, the accepted option keys in lexer and parser editors are
the same but this will be fixed shortly.