Adding content assist proposal for option partitions in lexer and parser editors
Only the options that are relevant to lexers (resp. parsers) are
proposed in lexer (resp. parser) editors.

Unfortunately, when an option is in the process of being written, it
does not yet match an option partition so it is not super useful as it
is. Maybe they should be proposed in the default partition as well,
depending on the context.