Replace FastPartitioner by a customized partitioner
This changes the default "optimized" behaviour of FastPartitioner which
often only repartitions a document from the beginning of the changed
line instead of rescanning from the beginning of a partition. For
complex partitions (such as semantic actions where the delimiters { }
can be nested) this leads to annoying behaviour in the editor when
editing in the middle of a partially complete partition.

The customized partitioner forces the partitioning to start from the
beginning of a partition, doing potentially more work but helping
partitions to remain consistent more often.

NB: It may be that the defautl damager/repairer in the reconciler has a
similar optimization because even when partitions remain correct or so
at all time, syntax highlighting remains wrong sometimes. This should be
looked into.