Adding JDT-style occurrence marking to Dolmen editors
This is a first attempt at the feature, the basic functionality is there
but there are many rough edges to polish:
- cleaning/updating of annotations on window (de)activation
- the resolving of references could be optimized
- some fragile behaviour regarding selections in semantic actions
- occurrence marking in lexer editors is almost pointless because
references to entries are hidden in semantic actions, and references to
defined regexps are inlined in the AST
- documentation..

Lexer editors can mark regexps and lexer entries, whereas parser editors
can mark terminals and non-terminals.
This reuses the annotation types that the JDT uses ("write occurrence"
for a declaration, "read occurrence" for references), because the
default annotation model works with these and also simply by uniformity
with what Java developers are used to.