Reintroduce literal partitions in .jl documents
Revert removal of literal partitions from lexer descriptions which was
performed in commit 5e86ff288dc85d159e9b006ebf3dfd1195126adf.

This is mandatory because special sequences such as //, /* or { } which
serve to identify other partitions can appear in literals and must not
be mistaken for these partitions. For comments, we could stop using
partitions and move their recognition to the scanning rules but Java
partitions are required to have dedicated behaviour of content assist,
hovers, etc in semantic actions.