More robust partitioning in both lexer and parser editors
The former customization of FastPartitioner we had worked ok because
there were no more literal partitions, so for instance an incomplete
Java partition was most of the time in the default partition and the
strategy to rescan from the start of the previous partition was fine.

With literal partitions in a mix, literals in incomplete semantic
actions appear as partition themselves and a more radical approach is
needed to keep the partitions in sync. We now rescan _from the start_ if
the document change contains characters which can lead to big changes in
partitioning (typically escape characters or opening/closing sequences
of partitions). I have tested that this performed reasonably well still
on a 1kloc lexer description. Parsers can get much bigger than that
though, so probably there will be some limit as to how far we are ready
to rescan (or maybe do it every 10 document changes or something like
that?).