Adding recognition of holes in extents to Java partition scanners
This is optional to account both for lexers (where extents are not
parametric) and grammars (where extents can now have holes, at least
some of them).

Note that the partitioning & highlighting is slightly more permissive
than the syntax, as for instance header/footer/token values can be
scanned with holes although they would not be syntactically correct in
Dolmen parsers.