Adapted annotation hovers to parametric grammars
The same declaration/references as before are supported for now.
In particular, we can't mark occurrences (or jump to the declaration) of
a formal parameter yet.
The rules' formal parameters, if any, are displayed in the hovers but it
would be nice to add the inferred sorts for each formals.