Adapting problem markers to source mappings with composite extents
Refined origin of source mappings can help locate Java problems back to
the adequate semantic action, even if it was the result of
instantiations.

Basic functionality is operational, but this requires polishing. In
particular, we should try and merge similar Java problems (i.e. Java
problems which arise at most instantiations of some rule/action) to
avoid reporting all of them back on the same spot as different markers.
Also, textual replacements can contain material which should be escaped
in HTML text hovers (although it seems to be fine in annotation hovers).