Forward JDT markers that relate to the same origin as a single Dolmen marker
Instead of reporting several near-identical markers on the same
position, we record a single marker and keep track of all expanded rules
which have been producing markers from that position. A single rule can
produce more than one marker from the same position, typically within
the result type action of the rule, as it is used more than once in the
generation of the rule.
We show the most recurring one in the Dolmen marker message.

NB: Messages are raw text and can contain HTML-specific symbols which
will confuse the fancy text hover control we use. On the other hand,
using HTML does confuse the non-fancy annotation hover control we
use, as well as the problem view. This must be fixed.